trezor/keepkey: merge qt_generic.py into qt.py
(remnants of separating trezor and keepkey)
